improved callout box styling (#77)
This commit adds a new shortcode, admonition, to replace notice
and alert shortcodes. It has nicer stylings, pulled from
mkdocs-material, giving the box a header and border.

Example: #+attr_shortcode: tip "This is a tip" followed by
a #+begin_admonition. The string title is optional, the
default would be the name of the styling, in this case "Tip".