Introduce Erlang LS (#93)
The current documentation does not mention Erlang LS as a possibility
in the "Setup IDE" section. Given the current progress of Erlang LS, I
think it's worth mentioning. I am also removing the reference to
`sourcer` for a couple of reasons:

* it's still experimental (like already stated in the same section)
* it offers a subset of the Erlang LS features

I have also moved the mentioning of language servers outside of the VS
Code section since, by definition, language servers are editor-agnostic.

If you would prefer to keep the reference to Sourcer I'm happy to do
so, please advise.